+++ ooniprobe/Utility/LocalizationUtility.mm
@@ -62,7 +62,7 @@
     return settingName;
 }
 
-+(NSString*)getSingularPlural:(int)value :(NSString*)locString{
++(NSString*)getSingularPlural:(long)value :(NSString*)locString{
     NSString *string;
     if (value == 1){
         string = [NSString stringWithFormat:@"%@.Singular", locString];
